Moved entity class down to server folder
