+++ src/gameserver/systems/player.ts
@@ -1,4 +1,4 @@
-import { Entity } from "../states/gameplay/entity";
+import { Entity } from "../server/entity";
 import { BaseState } from "../server/basestate";
 
 export function playerSystem(ents: ReadonlyArray<Entity>, state: BaseState) {
